+++ EFREI/Conception_Back_End/minitp/src/controllers/TrainerController.ts
@@ -1,5 +1,6 @@
 import TrainerService from '@/services/TrainerService';
 import { Request, Response } from 'express';
+import { validationResult } from 'express-validator';
 
 class TrainerController {
   constructor(private trainerService: TrainerService) {}
@@ -17,6 +18,12 @@ class TrainerController {
 
   /** GET /trainers/:id */
   public getById = async (req: Request, res: Response): Promise<void> => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() });
+      return;
+    }
+
     const trainer = await this.trainerService.getTrainerById(parseInt(req.params.id));
 
     if (!trainer) {
@@ -33,12 +40,24 @@ class TrainerController {
 
   /** POST /trainers/:id/heal - Heals all trainer's Pokémon */
   public healAll = async (req: Request, res: Response): Promise<void> => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() });
+      return;
+    }
+
     await this.trainerService.healAllPokemons(parseInt(req.params.id));
     res.json({ message: 'Tous les Pokémon ont été soignés' });
   };
 
   /** POST /trainers */
   public create = async (req: Request, res: Response): Promise<void> => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() });
+      return;
+    }
+
     const { name } = req.body;
     const trainer = await this.trainerService.createTrainer(name);
     res.status(201).json(trainer);
@@ -46,6 +65,12 @@ class TrainerController {
 
   /** DELETE /trainers/:id */
   public delete = async (req: Request, res: Response): Promise<void> => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() });
+      return;
+    }
+
     await this.trainerService.deleteTrainer(parseInt(req.params.id));
     res.status(204).send();
   };
